message_relay/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
cfgv==3.4.0
    # via pre-commit
confluent-kafka==2.3.0
    # via -r requirements.in
distlib==0.3.7
    # via virtualenv
filelock==3.13.1
    # via virtualenv
identify==2.5.32
    # via pre-commit
nodeenv==1.8.0
    # via pre-commit
platformdirs==4.0.0
    # via virtualenv
pre-commit==3.5.0
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pyyaml==6.0.1
    # via pre-commit
structlog==23.2.0
    # via -r requirements.in
virtualenv==20.24.7
    # via pre-commit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
